# src/main/java/io/swagger/api/CartApi.java
/**
 * NOTE: This class is auto generated by the swagger code generator program (3.0.46).
 * https://github.com/swagger-api/swagger-codegen
 * Do not edit the class manually.
 */
package io.swagger.api;

import io.swagger.model.Cart;
import io.swagger.v3.oas.annotations.Operation;
import io.swagger.v3.oas.annotations.Parameter;
import io.swagger.v3.oas.annotations.enums.ParameterIn;
import io.swagger.v3.oas.annotations.responses.ApiResponses;
import io.swagger.v3.oas.annotations.responses.ApiResponse;
import io.swagger.v3.oas.annotations.media.ArraySchema;
import io.swagger.v3.oas.annotations.media.Content;
import io.swagger.v3.oas.annotations.media.Schema;
import io.swagger.v3.oas.annotations.security.SecurityRequirement;
import org.springframework.http.ResponseEntity;
import org.springframework.validation.annotation.Validated;
import org.springframework.web.bind.annotation.PathVariable;
import org.springframework.web.bind.annotation.RequestBody;
import org.springframework.web.bind.annotation.RequestHeader;
import org.springframework.web.bind.annotation.RequestMapping;
import org.springframework.web.bind.annotation.RequestMethod;
import org.springframework.web.bind.annotation.RequestParam;
import org.springframework.web.bind.annotation.RequestPart;
import org.springframework.web.multipart.MultipartFile;
import org.springframework.web.bind.annotation.CookieValue;

import javax.validation.Valid;
import javax.validation.constraints.*;
import java.util.List;
import java.util.Map;

@javax.annotation.Generated(value = "io.swagger.codegen.v3.generators.java.SpringCodegen", date = "2023-09-28T16:59:58.194634916Z[GMT]")
@Validated
public interface CartApi {

    @Operation(summary = "Artikel aus dem Einkaufswagen entfernen", description = "", tags={ "cart" })
    @ApiResponses(value = { 
        @ApiResponse(responseCode = "200", description = "Artikel erfolgreich aus dem Einkaufswagen entfernt"),
        
        @ApiResponse(responseCode = "400", description = "Ungültige Anforderung") })
    @RequestMapping(value = "/cart/{product_id}",
        method = RequestMethod.DELETE)
    ResponseEntity<Void> cartProductIdDelete();


    @Operation(summary = "Artikel zum Einkaufswagen hinzufügen", description = "", tags={ "cart" })
    @ApiResponses(value = { 
        @ApiResponse(responseCode = "200", description = "Artikel erfolgreich zum Einkaufswagen hinzugefügt"),
        
        @ApiResponse(responseCode = "400", description = "Ungültige Anforderung") })
    @RequestMapping(value = "/cart/{product_id}",
        consumes = { "application/json" }, 
        method = RequestMethod.POST)
    ResponseEntity<Void> cartProductIdPost(@Parameter(in = ParameterIn.DEFAULT, description = "", required=true, schema=@Schema()) @Valid @RequestBody Cart body);

}