# src/main/resources/openapi3_0.yaml
openapi: 3.0.0
info:
  title: Lebensmittel-Shop API
  version: 1.0.0
paths:
...
  /cart/{product_id}:
    post:
      tags: 
        - cart
      summary: Artikel zum Einkaufswagen hinzufügen
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/Cart"
      responses:
        '200':
          description: Artikel erfolgreich zum Einkaufswagen hinzugefügt
        '400':
          description: Ungültige Anforderung
    delete:
      tags:
        - cart 
      summary: Artikel aus dem Einkaufswagen entfernen
      parameters:
        - name: product_id
          schema: 
            $ref: '#/components/schemas/Cart'
        
      responses:   
        '200':
          description: Artikel erfolgreich aus dem Einkaufswagen entfernt
        '400':
          description: Ungültige Anforderung     
...
  /address:
    post:
      tags:
        - address
      summary: Neue Adresse hinzufügen
      description: Neue Adresse hinzufügen
      operationId: addAdress
      requestBody:
        description: Neue Adresse hinzufügen
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Address'
          application/xml:
            schema:
              $ref: '#/components/schemas/Address'
          application/x-www-form-urlencoded:
            schema:
              $ref: '#/components/schemas/Address'
        required: true
      responses:
        '200':
          description: Erfolgreich hinzufügt
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Address'      
            application/xml:
              schema:
                $ref: '#/components/schemas/Address'
        '405':
          description: Ungültiger Input
    get:
      tags:
        - address
      summary: Alle Addressen finden
      description: 'Gibt ein Array mit allen Adressen aus'
      operationId: findAllAddresses
      responses:
        '200':
          description: Erfolgreich gefunden
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Address'       
            application/xml:
              schema:
                $ref: '#/components/schemas/Address'
  /address/{address_id}:
    put:
      tags:
        - address
      summary: Adresse updaten
      description: Adresse updaten
      operationId: updateAddress
      requestBody:
        description: Adresse updaten
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Address'
          application/xml:
            schema:
              $ref: '#/components/schemas/Address'
          application/x-www-form-urlencoded:
            schema:
              $ref: '#/components/schemas/Address'
        required: true
      responses:
        '200':
          description: Update erfolgreich
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Address'        
            application/xml:
              schema:
                $ref: '#/components/schemas/Address'
        '400':
          description: Ungültige ID
        '404':
          description: Addresse nicht gefunden
        '405':
          description: Validation exception
       
    delete:
      tags:
        - address
      summary: Adresse löschen
      description: ''
      operationId: deleteAddress
      parameters:
        - name: product_id
          in: path
          description: Adress id to delete
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '400':
          description: Ungültige ID
    get:
      tags:
        - address
      summary: Adresse finden
      description: ''
      operationId: findAdress
      responses:
        '200':
          description: Erfolgreich gefunden
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Address'        
            application/xml:
              schema:
                $ref: '#/components/schemas/Address'
        '400':
          description: Ungültige ID
        '404':
          description: Adresse nicht gefunden
components:
  schemas:
...
    Cart:
      type: object
      properties:
        cart_id:
          type: integer
        product_id:
          type: string
        quantity:
          type: integer
        total:
          type: integer
...
    Address:
      type: object
      properties:
        address_id:
          type: integer
        street:
          type: string
          example: 437 Lytton
        city:
          type: string
          example: Palo Alto
        state:
          type: string
          example: CA
        zip:
          type: string
          example: '94301'
      xml:
        name: address
...